backend/agents/demand_forecast.py
from typing import Dict, Any
from backend.agents.audit import log_agent_workflow
from langsmith import traceable

@traceable(run_type="chain", name="Demand Forecast Agent")
def forecast_demand(shop_id: str, historical_days: int = 30) -> Dict[str, Any]:
    """
    Agent stub: Analyzes shop historical data and tourism trends to predict inventory needs.
    """
    input_data = {
        "shop_id": shop_id,
        "historical_days": historical_days
    }
    
    decision = {"restock_recommendations": [{"product_id": "uuid-placeholder", "quantity": 50}]}
    confidence = 0.72
    
    log_agent_workflow("demand_forecast", input_data, decision, confidence)
    return decision
