pyproject.toml
[tool.poetry]
name = "dhivashi-backend"
version = "0.1.0"
description = "Dhivashi Phase 0 MVP Backend - FastAPI & LangGraph"
authors = ["Dhivashi <admin@dhivashi.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
supabase = "^2.4.5"
langgraph = "^0.0.30"
langchain = "^0.1.13"
langchain-openai = "^0.1.1"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
